pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 80
target-version = ['py310']
skip-string-normalization = true

[tool.isort]
line_length = 80
profile = "black"

[tool.mypy]
python_version = '3.10'
pretty = true
incremental = false
disallow_any_unimported = true
disallow_any_expr = true
disallow_any_decorated = true
disallow_any_explicit = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true
allow_untyped_globals = false
allow_redefinition = false
strict = true
strict_equality = true
show_error_codes = true
disable_error_code = [
    'misc',
    'no-any-unimported',
    'no-any-return',
    'union-attr',
]
exclude = [
    'venv/',
]

[[tool.mypy.overrides]]
module = [
    "notifypy",
    "watchdog.*"
]
ignore_missing_imports = true

[tool.poetry]
name = "pycnysr"
version = "0.0.1"
description = "a simple multi directory watcher and syncer"
authors = ["Laurent Baillet <laurent.baillet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pydantic = "^1.9.1"
pyyaml = "^6.0"
temppathlib = "^1.2.0"
watchdog = "^2.1.9"
types-pyyaml = "^6.0.12"
coverage = {extras = ["toml"], version = "^6.5.0"}
pytest-cov = "^4.0.0"
click = "^8.1.3"
loguru = ">=0.5.3,<0.6.0"
notify-py = ">=0.3.3,<0.4.0"

[tool.poetry.group.dev.dependencies]
autoflake = "^1.7.5"
flake8-markdown = "^0.4.0"
isort = "^5.10.1"
mypy = "^0.982"
pytest = "^7.1.3"
yamllint = "^1.28.0"
toml-sort = "^0.20.1"
pre-commit-hooks = "^4.3.0"
pre-commit = "^2.20.0"
black = "^22.10.0"

[tool.poetry.scripts]
pycnysr = 'pycnysr:main'
